requirements.txt
-f /usr/share/pip-wheels
Babel==2.8.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1
django-bootstrap4==1.1.1
django-braces==1.14.0
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-dajax==0.9.2
django-dajaxice==0.7
django-embed-video==1.3.3
django-heroku==0.3.1
django-js-asset==1.2.2
django-memcache-status==1.1
django-phonenumber-field==4.0.0
django-taggit==1.3.0
djangorestframework==3.11.1
gunicorn==20.0.4
idna==2.9
oauthlib==3.1.0
phonenumbers==8.12.7
Pillow==7.1.1
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
python3-memcached==1.51
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==2.1.0
social-auth-core==3.3.3
soupsieve==2.0
urllib3==1.25.8
validate-email==1.3
whitenoise==5.2.0
